Bug#16665072: REMOVE BOGUS ERRNO DECLARATION
This patch removes "extern int errno" and instead unconditionally
does "#include <errno.h>". According to the man page:

"It  was common in traditional C to declare errno manually (i.e., extern
int errno) instead of including <errno.h>.  Do not do  this. It will
not work with modern versions of the C library."

This also removes the need for manually patching rpm-uln and
Fedora builds to fix this issue.