[t:5153] fix the error handling code for toku_ltm_get_lt() and better explain what the client contract is for the on_create and on_close callbacks.
the idea is you should only pass an on_create callback if there is also an on_close callback, and vice-versa. the on_create callback is called if
the lock tree is created and the on_close callback is called only if the on_create callback was actually called, either because of a user close
of the locktree or by an error path in toku_ltm_get_lt().

we never assume that either of these callbacks are non-null and we know both can be called at most once.


git-svn-id: file:///svn/toku/tokudb@45002 c7de825b-a66e-492c-adef-691d508d4ae1