WL#4772: Improve building of classes that represent DML statements
The classes st_select_lex_unit and st_select_lex are now always
allocated dynamically: Before this change, objects for the top-level
query block was kept as member of the LEX class.

The changes made in bug#14828360, which removed the base class
st_select_lex_node, have also been incorporated in this patch.

Some refactoring as to how objects of these classes are created has
been done. These functions are collected in the LEX class.

Proper constructors have been added for the above-mentioned classes,
which should reduce probability for mistakes when re-initializing
query objects and hence reduce maintenance costs.

Numbering of query blocks and marking of level of query blocks in
query tree has been simplified. As a consequence, query block numbers
has changed somewhat. This is one of the few user-visible impacts
of this work.

Handling of statement options that only apply to the parsing stage
has been moved from select_lex and into the class Parser_state.