Bug#16267120 - PAGE REORGANIZE ASSUMES INNODB_LOG_COMPRESSED_PAGES=OFF
btr_page_reorganize() reorganizes a B-tree page to get more
available space for inserting a record.

For compressed pages, this is being logged as a logical operation. This
is wrong, because the zlib version or the compression parameters could
change between the time the MLOG_COMP_PAGE_REORGANIZE redo log record is
written and it is applied during a crash recovery.

MySQL 5.6 introduces the parameter innodb_compression_level and 
innodb_log_compressed_pages.

The problem is that page reorganize is always being logged as
MLOG_ZIP_PAGE_REORGANIZE (with the compression level as a parameter),
which is correct with innodb_log_compressed_pages=OFF.

If innodb_log_compressed_pages=ON, page reorganize should log the
compressed page image.

Fix:
---
If innodb_log_compressed_pages=ON, log the compressed page and do not
write a MLOG_COMP_PAGE_REORGANIZE redolog record.

Approved by Marko. rb#1974