Bug#16684523 INFORMATION_SCHEMA.INNODB_LOCKS UNNECESSARILY FETCHES BLOCK FOR INFIMUM,SUPREMUM
The function fill_locks_row(), which is responsible for providing the
data for the INFORMATION_SCHEMA.INNODB_LOCKS table, tries to look up
the B-tree page in the buffer pool even for the pre-determined infimum
and supremum records, which have pre-determined heap_no 0 and 1. This
will generate some unnecessary contention on the buffer pool, and will
unnecessarily omit some information when the page does not happen to
be available in the buffer pool.

Approved by Sunny Bains