Bug#16338667 - DROP DATABASE FAILS ON EXTERNAL DATA DIRECTORY
MySQL cleans up the database directory if there are only files that
it recognizes.  If there are unknown file types, the DROP DATABASE
will fail.  We forgot to tell the server in ha_innobase::bas_ext()
that one if the possible extentions of InnoDB files is ".isl".

Approved by Marco in http://rb.no.oracle.com/rb/r/2337/