Bug#16403420-ROW FORMAT CHANGING ALTER TABLE FAILS IN              COMMIT_CACHE_REBUILD, RENAMING
This bug was opened when to a temporary tablespace pre-existed with
the same name during an ALTER TABLE. This patch does not attempt to
discover or fix the reason that an orphan temporary tablespace existed.
Instead it tries to eliminate the chance of that occurring again by
making the temporary table/tablespace name effectively unique for any
MySQL database instance.

It does this by adding the current LSN at the end of the previous
table/file name.  For example; tablename "test/#sql-ib21" becomes
something like "test/#sql-ib21-1701208" where 1701208 is the
current_lsn.  Both the lsn and the table ID are needed to assure that
the name is unique because it is theoretically possible for multiple
threads to get the same LSN, but they will be working on differrent
tables.  Also, it is still possible to be able to relate the temporary
name to the table by the table ID.

Unfortunately, there are some recent testcases that failed because
they assume to know what the name of a temporary tablename/tablespace
name will be. One of them ia removed here and the other two are
salvaged by reading the actual assigned name from
information_schema.innodb_sys_tables.

Approved by Marko in http://rb.no.oracle.com/rb/r/2306/