Bug#16579028 	"SEMI-SYNC MASTER WAIT FOR REPLY FAIL TO GET WAIT TIME" HAPPENS OCCASIONALLY
The error of 'Semi-sync master wait for reply fail to get wait time'
happens if the time taken by dump thread on waiting for binlog reply
from slave is less than zero.
  
The precision of OS timer apparently causes this problem. But this
problem will not affect any function of the server, so it's not
good to print an error into the log file as the error will confuse
the user. It's more appropriate to print a note into the log file.