Bug#12772601 CENTROID FUNCTION ASSERTS: ASSERTION FAILED: N_LINEAR_RINGS > 0 Bug#16510712 SERVER CRASHES IN GET_POINT ON A GEOMETRY QUERY
Problem: most of the Geometry methods that work with WKB
did not properly validate input data, so they could
access memory beyond the range of provided data, which
led to crashes and/or valgrind errors.

Fix: introducing new class wkb_parser with methods
that strictly validate the provided WKB data;
convenience changes and clean-ups in the related code.