Fix for bug#16678033:
  IN-SUBQUERY WITH TRUE/FALSE/UNKNOWN TEST MAY RETURN INCORRECT RESULT OR ASSERT

We failed to set an endpoint in ha_ndbcluster::index_read() when 
it was used to read an EQ_RANGE.

When used to read from an 'UNIQUE HASH INDEX', *and* the
key contained NULL values, this will be executed as a table scan
and the range will be used to construct a scan_filter.
The missing endpoint caused this scan to incorrectly
return the first row after the requested row if it didn't
exists.

This in turn caused subselect_indexsubquery_engine::exec() to
incorrectly conclude that it found a row with NULL valued key
and return incorrect results.