Bug #16604641: "SET PASSWORD FOR" CMD SUCCESS FOR USER PWD  EXPIRED & FLAG NOT SET ACCORDINGLY
The expired flag in the thread specific copy of the user information 
was frivolously reset without first checking if the user whose password
is updated is the same user as in the thread cache.

Fixed by properly checking if the thread cache needs to be updated too
(together with the global user cache and the disk version).

Also fixed the omision to enforce the password expiration mode on 
SET PASSWORD FOR ... statement for cases where the password for
the current user is not reset.

Test case added.