This is a patch for bug#16388996 - ASSERTION IN TRX_START_IF_NOT_STARTED_LOW                                    WITH QUERY CACHE ENABLED.
This bug relates to working with XA-transaction when query cache is enabled.
In some cases when XA-transaction is in a state either IDLE or PREPARED
the execution of QUERY with query_cache enable could led to a crash for the
server built in debug mode.

To fix this bug we added checking for a state of XA-transaction before
looking for the query in the cache. If a state of XA-transaction is an IDLE or
PREPARED the query cache isn't called to look for the result of query
execution.