Bug#16697792: POOR EXECUTION PLAN WHEN ORDER BY WITH LIMIT X
When we have a LIMIT that is lower than the row estimate on 
the first table, the optimizer checks whether or not it's 
possible to switch to an index that provides correct ordering.
If there is, execution can be short-cut'ed immediately after
finding 'limit' number of records. The optimizer (in 
make_join_select()) will in such cases go through all ORDER BY
items and make note of all indexes that can be used to 
provide ordering for that item and then run the range optimizer
to see if changing index makes sense.

That happens for this bug too. make_join_select() looks at 
the ORDER BY clause and finds indexes that cover the fields.
The range optimizer is then run to see if those indexes are
useful. In the bug, the indexes with that index number may not
be used, so range access is discarded. The problem is that the 
first table - the one we run range optimizer for - is not the 
same table referred to in the ORDER BY clause. Thus, the range 
optimizer will check if an index in a different table is 
useful.

The fix is to check if the fields in the ORDER BY clause are 
from the table being analyzed, and leave the query plan as 
earlier computed if any of them belong to a different table.