wl#6244 bugXXV This patch fixes a problem with TC triggers i.e unique indexes, table-reorg, foreign keys (and all nodes)
suppose you send a TCKEYREQ (ignore error)
and "works" on primary replica, fireing trigger
and there is an error on backup replica
the fired triggers are kep on transaction.
This patch changes this so that fired triggers are kept
on the operation until the operation gets LQHKEYCONF
and is only then moved onto the transaction list.