Bug#16437940: BAD CHANGE FROM REF TO RANGE DUE TO HEURISTICS
The MySQL optimizer has the following heuristic: "if decision 
is to use ref access on index 'i', and it is possible to use 
range access utilizing more keyparts than ref on the same 
index 'i' => switch to range access".

This heuristic makes sense when ref access refers to a 
constant, but is not so good when it refers to a column of a 
table earlier in the join sequence. In this case the range 
access method will not utilize the join condition.

Fix: don't apply the heuristic if 'ref' access refers to a 
column of another table.