WL#6871: Remove UNIV_BLOB_DEBUG.
UNIV_BLOB_DEBUG assumes that offsets[] be available, but we must try
to avoid computing them in vain.

If UNIV_BLOB_DEBUG is to be revived, it should be implemented
with std::map instead of the homebrew ut0rbt data structure,
and it should be written so that it tolerates crash recovery.