WL#6871: PageCur: Compute offsets on demand only.
The offsets are really only needed in PageCur::search() so far.  In
all other operations implemented so far, it suffices to know the size
of the record. (For inserting into compressed pages, we will need to
know offsets, so that some fields can be stored in uncompressed
format.)

PageCur::m_comp: Flag: is the record not in ROW_FORMAT=REDUNDANT?

PageCur::getRecSize(): New method for computing the size of the record.

PageCur::getOffsetsIfExist(): Get the m_offsets if they exist.

PageCur::getOffsets(): Compute the offsets if they do not exist.

rec_get_size_comp(): Add an optional output parameter n_ext.

PageCur::insert(): PageCur::purge(): Replace getOffsets() with getRecSize().