Move sql engine #include files to ha_prototypes.h
This is a non-functional patch that moves all #include references
to files outside InnoDB into a single file; ha_prototypes.h.
It collects all these includes into one place for better tracking.
The advantage is that this could become a precompiled header on some
compilers.  The disadvantage is that most changes to header files in
the server will cause most of InnoDB to recompile.

Grouping these headers together exposed a couple naming conflicts;

1) The InnoDB function ha_create() needed to be renamed to
ib_create() because it now conflicts with a handler.h function of
the same name.

2) The enum value "ERROR" in Group_cache::enum_add_group_status
needed to be renamed, according to an error in the VS2010 compile.

All dependencies from InnoDB into the server can now be found by
searching for ha_prototypes.h which is included in about 50 files.

Approved by Sunny in RB#2364