WL#6355 Semisync: externalize transactions only after ACK received
Postfix
There was a assert failure in Binlog_storage_delegate::after_sync.
It asserts that the argument log_pos is never 0. But It happened
if the last thd in the commit queue encountered an flush error.

To fix the problem, the log_pos of the last thd without error
is passed to after_sync.