BUG#16315351 - ALTER TABLE DROP COLUMN: NO WARNING IF DUPLICATE                INDEX FOR THE SAME COLUMN
Analysis
-------

When a column associated with a multi-column index is dropped,
a NOTE is not reported in case a duplicate index exist on the
table after the DROP COLUMN operation.

EX: CREATE TABLE t1( fld1 INT, fld2 INT, KEY(fld1, fld2),    
                     KEY(fld2));
    ALTER TABLE t1 DROP COLUMN fld1;

The duplicate index check was not invoked in the above mentioned
case.

Fix:
----
The flag 'check_for_duplicate_indexes' is set, to trigger the 
duplicate index check when the column associated with an index 
is dropped.