Bug #16241992
A COM_CHANGE_USER failure costs very little and
is not a subject to the same accounting a login failure
is. This creates an unfair advantage over the ordinary
login process.
Fixed by making COM_CHANGE_USER failing to login 
poison the connection  (using an unique error number)
and cause disptatch_command() to exit with an error
 instead of reverting back to the previous credentials.
Test cases updated.