BUG#16223835: UNEXPECTED EMPTY RESULTS OUT OF SHOW COMMANDS AFTER DUP GTID_NEXT TRANSACTION
When a transaction was skipped due to its GTID was already logged,
until GTID_NEXT was set to a different GTID, all the following
executed transactions were incorrectly skipped.

To avoid this incorrect behaviour, all executed transactions and
skipped transactions on AUTOCOMMIT=0 mode or which implicit commit
are marked as undefined (GTID_NEXT.type = UNDEFINED_GROUP) when them
commit or rollback. So when a second transaction is executed for the
same SET @@SESSION.GTID_NEXT command a error is thrown.