Fix for Bug#16744050 DUPLICATED ROWS RETURNED FROM IN-SUBQUERY
When ha_ndbcluster::position() copy VARCHAR fields into ref_key buffer,
only the actuall length of the VARCHAR fields should be copied and the 
rest of the VARCHAR zero-padded in order to create a predictable binary
representation of the position snapshot.

This is due to the MySQL Execution Engine, at least from ver5.6, assumes
that two position-instances of the same primary key are binary comparable.
Similar zero padding of the position snapshot is already present in
ha_innodb::position().

Only affect (AFAIK) execution correctness for versions >= 7.3
where it fails due to the new 'duplicate weedout' algorithm.
However, the fix is pushed from 7.0 -> as there is a bug
in ha_ndbcluster::position().

Testcase will be added as part of merge to 7.3.