Bug#69119 Bug#16760474 Wrong FOUND_ROWS() on MySQL 5.6.11 Bug#69144 Bug#16764220 sql_calc_found_rows, found_rows() in 5.6.11 differs from 5.6.10 Bug#69166 Bug#16775557 Wrong result of FOUND_ROWS() in MySQL 5.6
This is a regression caused by the patch for
Bug#68458 Wrong count from FOUND_ROWS() on MySQL 5.6
That patch restricted the setting of "thd->limit_found_rows"
to the case where filesort had actually used a Priority Queue.

The correct solution is to restrict the setting to the case
where filesort *could* have used a priority queue
(i.e. whenever we have "order by with limit")
See code in check_if_pq_applicable() which returns false
if it thinks that quicksort will be cheaper that PQ.