ndb - bug#16834416
This testcase shows a leak of commitackmarkers.
This problem can manifest it self as a "normal" leak,
i.e resource will sooner or later be exhausted.
It can also show it self as crash in TC or LQH due
to duplicate markers.

Add in_use flag to DataBuffer::Head
and change all occurances of in_use to also be enabled
with #ifdef ERROR_INSERT

This problem was introduced in
2012-10-18 - pekka.nousiainen@oracle.com-20121018065634-k4dcob7k4luvzhp5

--

Remove unneeded and harmful instantiations of LocalDataBuffer