Bug#16756402-RECENT REGRESSION: CRASH IN ITEM_DEFAULT_VALUE::WALK
Problem:-
It is a regression, cause by the patch of bug#14789787.

Analysis:
In Item_default_value::walk, while traversing in the argument 
of Item_default_value. We are not taking care whether 
Item_default_value has any argument or not.

Solution:
Before traversing, check whether it carry any argument or not.