Followup to Bug#16821155 ADD A CMAKE OPTION TO EASILY ENABLE EXTRA INNODB DEBUG CHECKS
In mem_heap_is_top() also check if the requested size of the buffer actually
matches the real size of the buffer (which is only recorded if UNIV_MEM_DEBUG
is defined). If the sizes do not match, then consider that the requested
buffer is not the top of the heap.

This is necessary because in mem_heap_free_top() we call mem_field_erase()
(in UNIV_MEM_DEBUG only) which asserts that the sizes match.