Followup to Bug#16821155 ADD A CMAKE OPTION TO EASILY ENABLE EXTRA INNODB DEBUG CHECKS
Relax a bit a too strict checks from innodb_cmp_per_index.test. It is not
the purpose of the test to check whether exactly 43 or 65 compression ops
occured. Thus change it to also accept the different numbers which are
returned when UNIV_ZIP_COPY and UNIV_ZIP_DEBUG are defined when
cmake -DWITH_INNODB_EXTRA_DEBUG:BOOL=ON is used.