Bug #16723708: INVALID VALUES RETURNED IN NDB_LOGEVENT "CATEGORY" MEMBER VARIABLE
For every log event retrieved from the MGM API, the log event category
was being cast to an enum type. This cast resulted in invalid category
values.

This was fixed by adding an offset to the category after casting to make
it match the valid enum values. A test case for this bug is added in 
testMgm.