ndb - testMgm
This patches "fixes" the stability problems of testMgm primarily on windows.
The root cause seems to be the testcase Bug45497, which will create connections to
ndb_mgmd until it takes more than 1s to connect. On windows this seems to create
some 1000 connections, and each connection is backed by one thread on ndb_mgmd.

This make the machines crawl...causing subsequent failures...
In fact I'm impressed it works so well as it does...

This patch introduces a testcase property "explicit",
which means that it will only be run called explicitly using name.
e.g "testMgm -n Bug45497" will run it, but "testMgm" will not.