Bug#11766880: CAN'T FIND FILE / FILE NAME TOO LONG
Overview
--------
The scenario here is identifiers, e.g., table names, with special
characters (e.g. newlines, ".", non-ascii, ...). When mapping
from table name to file name (for the .frm file), these special
characters are mapped to character sequences (e.g., '.' is mapped
to @002e). Hence, the file name may become much longer than the
table name, so the table name may very well be valid (i.e., <= 64
characters) while the file name is invalid (length > NAME_MAX,
e.g. 255 characters). E.g. a table name of 51 dots ("..." etc.)
is mapped to a 255 character file name. Thus, on trunk, when
attempting an insert into a non-existing table, we get the
following errors:

---------+---------+--------------------+--------------------
Table nm | File nm | OS Error (solx86)  | MySQL error
---------+---------+--------------------+--------------------
<=64     | <=255   | No such file       | Table doesn't exist
<=64     |  >255   | File name too long | Can't find file
 >64     | Any     | N/A                | Identifier too long
---------+---------+--------------------+--------------------

This behavior alone can hardly be considered a bug. However, it
is still relevant to consider potential buffer overflow issues.


Buffer usage and potential overflow issues
------------------------------------------
The same character mapping from db identifiers to filesystem
names is relevant also for schemas, meaning that we can get quite
long path names when concatenating the mapped names of a schema
and table with numerous special characters. When allocating the
table share while opening a table, in "build_table_filename", the
schema name and table name are converted and concatenated to form
a path name. This path name is put into a buffer of size
FN_REFLEN, which is 512 characters. Overflow is checked, and only
511 characters will be used, leaving the last character for the
'\0' termination of the string. If the file name is truncated,
this is indicated by a return parameter (which is unfortunately
ignored, see below).

When the table share structure is allocated, right afterwards,
space is allocated for the path name. The exact length (without
'\0') is allocated, plus one additional byte for the trailing
'\0'. The "path.str" and "normalized_path.str" data members in
the table share object are both set to point at the recently
allocated path array. So far so good.

But alas, in "open_table_def", the file name to open is formed by
concatenating the "normalized_path.str" from the table share with
the proper extension (e.g. ".frm"); this is squeezed into a
string with only FN_REFLEN bytes allocated, and hence, we have a
potential overflow issue: The path may be up to FN_REFLEN - 1
characters, and adding the ".frm" makes the string copy write
beyond the allocated size.


Ignored truncation of path
--------------------------
There is a potential issue related to the ignored truncation
after building the file name, which may in theory result in
different table names mapping to the same file name. However, in
practice, this seems to be hidden by the fact that the file names
will be too long anyway (path may be 511 characters, -2 for "./",
-255 (max) for the schema, -1 for the path separator, thus
leaving up to 253 characters for the file name part of the
path. Adding ".frm" makes this 257 characters (ignoring the
buffer overflow), which is > 255 and therefore will fail with
"file name too long" when opening the file, at least on solaris
11 x86 with max 255 characters per path component. And besides,
truncation seems to be checked when creating a table (in
"mysql_create_table_no_lock").


Creating tables
---------------
While investigating the overflow issues, code executed while
creating tables was also relevant to look at. Here, the following
scenario was found:

1. In "mysql_create_table_no_lock", FN_REFLEN + 1 bytes are allocated,
   and space for the ".frm" extension is checked.

2. After creating the frm file, the file was opened in
   "open_table_def". Here, however, only FN_REFLEN bytes were
   allocated, and with no length check, there was a potential
   overflow (same overflow as identified for inserts). 

3. Since "open_table_def" failed, the create statement execution
   failed and the file was to be deleted. In "rea_create_table",
   when concatenating the file name used, e.g., for delete, a
   buffer of only FN_REFLEN bytes was allocated and strings were
   concatenated without checking legth, so there would be a
   potential overflow.

Hence, the conclusion is that we should always allocate FN_REFLEN
+ 1 bytes to allow for paths up to FN_REFLEN bytes (accepted by
the OS) plus the trailing '\0', and we should always apply
strxnmov for length checking when concatenating strings.


Suggested fix
-------------
Some of the behavior described above would be expected to
disappear when the new data dictionary is implemented. The
suggested fix hence contains:

1. A set of tests illustrating the behavior described above.

2. A patch for the buffer overflow and ignored truncation in
   "open_table_def": Add one additional character to the buffer
   (to allow for 512 byte paths), subtract "reg_ext_length" from
   length parameter passed to "build_table_filename".

3. Rewrite "mysql_create_table_no_lock" slightly to use the same
   implementation pattern as "open_table_def" described above.
   
4. For an extra safeguard, replace "strxmov" by "strxnmov" when
   building the path with extension before opening the file in
   "open_table_def". Add one additional byte to the allocation
   for consistency with items above.

5. Similar to the item above, change the implementation in
   "rea_create_table": Add one byte when allocating the path, and
   check length when concatenating strings using strxnmov.